refactor client code
